A IR/Arduino goal line score system for a foosball table
System consists (or will do :) of:

A) A re-purposed Tilda E badge from EMFCamp 2014
https://badge.emfcamp.org/wiki/TiLDA_MKe
This is a Arduino Due with a small screen, battery and an SRF radio. I
could have used a Ciseco RFU328 but this was handy and came with
battery/screen etc. It shows scores locally and also sends them via SRF
Radio to
B) A RPi. This is because I have to have a wired LAN connection. The PI
sends out the scores to iot.eclipse.org via MQTT.
C) Something interesting to pick up the scores, send them to twitter
etc in true IoT style.